# api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Configure and inspect a running TDNS server.
  title: TDNS Management API
  version: "1.0"
paths:
...
  /api/zen-mode:
    get:
      description: Return zen mode configuration, runtime state, and timing.
      operationId: zenModeStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Get zen mode status
      tags:
        - zen-mode
    post:
      description: Replace the in-memory domains blocked by zen mode.
      operationId: zenModeDomainsReplace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ZenReplaceRequest"
        description: Zen mode domains
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace runtime zen domains
      tags:
        - zen-mode
  /api/zen-mode/persisted/domains:
    post:
      description: Replace zen domains stored in configuration overrides.
      operationId: zenModePersistedDomainsReplace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ZenReplaceRequest"
        description: Persisted zen mode domains
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace persisted zen domains
      tags:
        - zen-mode
  /api/zen-mode/persisted/excludes:
    post:
      description: Replace zen exclusions stored in configuration overrides.
      operationId: zenModePersistedExcludesReplace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ZenExcludesRequest"
        description: Persisted zen mode exclusions
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace persisted zen exclusions
      tags:
        - zen-mode
  /api/zen-mode/start:
    post:
      description: Start a zen mode period using the current configuration.
      operationId: zenModeStart
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Start zen mode
      tags:
        - zen-mode
...
servers:
  - url: /
components:
...
  securitySchemes:
    BearerAuth:
      description: Bearer token using the format "Bearer {token}".
      in: header
      name: Authorization
      type: apiKey
    CookieAuth:
      description: Opaque HttpOnly browser session cookie set by POST /api/auth/exchange.
      in: cookie
      name: __Host-tdns-session
      type: apiKey
  schemas:
...
    api.BlacklistStatus:
      properties:
        blockfile_total_entries:
          type: integer
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        external_file:
          type: string
        external_pull_period:
          type: string
        external_repo:
          type: string
        external_repo_branch:
          type: string
        file:
          type: string
        persisted_excludes:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            type: string
          type: array
        runtime_whitelist:
          items:
            type: string
          type: array
      type: object
...
    api.CacheStatus:
      properties:
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        hits:
          type: integer
        misses:
          type: integer
        ttl:
          type: integer
      type: object
    api.ClientCandidate:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
...
    api.DashboardHourlyPoint:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        hour_bucket:
          type: integer
        hour_start:
          type: string
        total_queries:
          type: integer
      type: object
    api.DashboardSummary:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        cache_hits:
          type: integer
        cache_misses:
          type: integer
        total_queries:
          type: integer
      type: object
...
    api.HostEntry:
      properties:
        address:
          example: 192.0.2.10
          type: string
        domain:
          example: example.com
          type: string
      type: object
    api.KnownHost:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
    api.LogDetails:
      properties:
        counter:
          example: 42
          minimum: 0
          type: integer
        domain:
          example: example.com.
          type: string
        host:
          example: office
          type: string
      type: object
...
    api.Response:
      properties:
        blacklist:
          $ref: "#/components/schemas/api.BlacklistStatus"
        cache:
          $ref: "#/components/schemas/api.CacheStatus"
        clients:
          items:
            $ref: "#/components/schemas/api.ClientCandidate"
          type: array
        current_status:
          example: enabled
          type: string
        hourly:
          items:
            $ref: "#/components/schemas/api.DashboardHourlyPoint"
          type: array
        items:
          items:
            type: string
          type: array
        kind:
          example: api.tdns/cache/response
          type: string
        known_hosts:
          items:
            $ref: "#/components/schemas/api.KnownHost"
          type: array
        log_items:
          items:
            $ref: "#/components/schemas/api.LogDetails"
          type: array
        message:
          example: Status OK
          type: string
        static_response:
          $ref: "#/components/schemas/api.StaticResponseStatus"
        stub_resolver:
          $ref: "#/components/schemas/api.StubResolverStatus"
        summary:
          $ref: "#/components/schemas/api.DashboardSummary"
        tag_members:
          items:
            $ref: "#/components/schemas/api.TagMember"
          type: array
        window_hours:
          maximum: 336
          minimum: 1
          type: integer
        zen_mode:
          $ref: "#/components/schemas/api.ZenModeStatus"
      type: object
...
    api.StaticResponseStatus:
      properties:
        configured_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        enabled:
          type: boolean
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        runtime_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
      type: object
...
    api.StubResolverStatus:
      properties:
        configured_stubs:
          items:
            type: string
          type: array
        enabled:
          type: boolean
        runtime_stubs:
          items:
            type: string
          type: array
      type: object
    api.TagMember:
      properties:
        address:
          example: 192.0.2.10
          type: string
        has_host_alias:
          type: boolean
        host:
          example: office
          type: string
      type: object
    api.ZenExcludesRequest:
      properties:
        excludes:
          items:
            type: string
          type: array
      type: object
    api.ZenModeStatus:
      properties:
        configured_domains:
          items:
            type: string
          type: array
        configured_excludes:
          items:
            type: string
          type: array
        duration_minutes:
          type: integer
        enabled:
          type: boolean
        ends_at:
          type: string
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_domains:
          items:
            type: string
          type: array
        persisted_excludes:
          items:
            type: string
          type: array
        remaining_seconds:
          type: integer
        runtime_domains:
          items:
            type: string
          type: array
        started_at:
          type: string
      type: object
    api.ZenReplaceRequest:
      properties:
        zen_domains:
          items:
            type: string
          type: array
      type: object